Powdery mildew disease progress curves in organic tomato

Experimental disease progress curve data for powdery mildew under different irrigation systems and soil moisture levels in organic tomato.

Format

A data frame with 240 rows and 5 variables:

irrigation_type

Irrigation system.

moisture

Soil moisture level.

block

Experimental block.

time

Assessment time.

sev

Disease severity as a proportion.

References

Lage, D. A. C., Marouelli, W. A., and Cafe-Filho, A. C. (2019). Management of powdery mildew and behaviour of late blight under different irrigation configurations in organic tomato. Crop Protection, 125, 104886.

Examples

data("PowderyMildew")
str(PowderyMildew)
#> 'data.frame':	240 obs. of  5 variables:
#>  $ irrigation_type: Factor w/ 5 levels "Drip","Furrow",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ moisture       : Factor w/ 2 levels "High moisture",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ block          : int  1 2 3 1 2 3 1 2 3 1 ...
#>  $ time           : int  0 0 0 7 7 7 14 14 14 21 ...
#>  $ sev            : num  0.011 0.011 0.011 0.034 0.04 0.036 0.114 0.109 0.097 0.126 ...