Fit models to multiple disease progress curves

Apply `fit_lin()`, `fit_nlin()`, or `fit_nlin2()` to multiple disease progress curves stored in a data frame.

Usage

fit_multi(
  time_col,
  intensity_col,
  data,
  strata_cols = NULL,
  starting_par = list(y0 = 0.01, r = 0.03, K = 0.8),
  maxiter = 500,
  nlin = FALSE,
  estimate_K = FALSE
)

Arguments

time_col

Character name specifying the time column.

intensity_col

Character name specifying the disease intensity column.

data

A data frame containing the variables for model fitting.

strata_cols

Character vector specifying grouping columns. Use `NULL` to fit all rows as a single epidemic. Defaults to `NULL`.

starting_par

Named list of starting values for model parameters.

maxiter

Maximum number of iterations for nonlinear fitting. Must be a positive number.

nlin

Logical. Should nonlinear fitting be used?

estimate_K

Logical. Should the asymptote `K` be estimated?

Value

A list with grouped parameter estimates and prediction data.

Examples

set.seed(1)
epi1 <- sim_gompertz(N = 30, y0 = 0.01, dt = 5, r = 0.3, alpha = 0.2, n = 2)
epi2 <- sim_gompertz(N = 30, y0 = 0.01, dt = 5, r = 0.2, alpha = 0.2, n = 2)
data <- dplyr::bind_rows(epi1, epi2, .id = "curve")
fit_multi(time_col = "time", intensity_col = "random_y", data = data, strata_cols = "curve")
#> $Parameters
#>   curve best_model         model         r        r_se   r_ci_lwr  r_ci_upr
#> 1     1          1      Gompertz 0.2954597 0.005145698 0.28424814 0.3066712
#> 2     1          2 Monomolecular 0.2555731 0.009434191 0.23501777 0.2761284
#> 3     1          3      Logistic 0.3684716 0.022879886 0.31862056 0.4183225
#> 4     1          4   Exponential 0.1128984 0.030578030 0.04627464 0.1795223
#> 5     2          1      Gompertz 0.2008236 0.004042745 0.19201523 0.2096320
#> 6     2          2 Monomolecular 0.1564261 0.008372027 0.13818504 0.1746672
#> 7     2          3      Logistic 0.2795232 0.017683136 0.24099492 0.3180514
#> 8     2          4   Exponential 0.1230970 0.024755054 0.06916042 0.1770337
#>           v0      v0_se r_squared       RSE       CCC          y0    y0_ci_lwr
#> 1 -1.4405713 0.09276540 0.9963734 0.1925344 0.9981834  0.01465304  0.005689949
#> 2 -0.5391255 0.17007730 0.9839115 0.3529951 0.9918905 -0.71450693 -1.483557804
#> 3 -3.0640435 0.41247302 0.9557780 0.8560870 0.9773891  0.04461503  0.018656188
#> 4 -2.5249180 0.55125328 0.5318337 1.1441251 0.6943752  0.08006488  0.024089103
#> 5 -1.5189406 0.07288163 0.9951605 0.1512657 0.9975744  0.01038509  0.004731929
#> 6 -0.4632800 0.15092887 0.9667688 0.3132526 0.9831036 -0.58927829 -1.208085307
#> 7 -3.3337389 0.31878726 0.9541760 0.6616423 0.9765507  0.03443171  0.017492827
#> 8 -2.8704589 0.44627808 0.6732629 0.9262493 0.8047306  0.05667291  0.021433175
#>     y0_ci_upr
#> 1  0.03173703
#> 2 -0.18359798
#> 3  0.10290625
#> 4  0.26611141
#> 5  0.02030762
#> 6 -0.14388944
#> 7  0.06666021
#> 8  0.14985269
#> 
#> $Data
#>     curve time          y         model    linearized   predicted      residual
#> 1       1    0 0.01000000   Exponential -4.6051701860  0.08006488 -7.006488e-02
#> 2       1    0 0.01000000 Monomolecular  0.0100503359 -0.71450693  7.245069e-01
#> 3       1    0 0.01000000      Logistic -4.5951198501  0.04461503 -3.461503e-02
#> 4       1    0 0.01000000      Gompertz -1.5271796258  0.01465304 -4.653038e-03
#> 5       1    5 0.36633668   Exponential -1.0042024658  0.14079847  2.255382e-01
#> 6       1    5 0.36633668 Monomolecular  0.4562375135  0.52228454 -1.559479e-01
#> 7       1    5 0.36633668      Logistic -0.5479649523  0.22763993  1.386968e-01
#> 8       1    5 0.36633668      Gompertz -0.0041936601  0.38138754 -1.505086e-02
#> 9       1   10 0.76788928   Exponential -0.2641097288  0.24760180  5.202875e-01
#> 10      1   10 0.76788928 Monomolecular  1.4605407604  0.86689347 -9.900420e-02
#> 11      1   10 0.76788928      Logistic  1.1964310316  0.65037138  1.175179e-01
#> 12      1   10 0.76788928      Gompertz  1.3313906230  0.80250071 -3.461143e-02
#> 13      1   15 0.96524451   Exponential -0.0353738300  0.43542129  5.298232e-01
#> 14      1   15 0.96524451 Monomolecular  3.3594177769  0.96291234  2.332170e-03
#> 15      1   15 0.96524451      Logistic  3.3240439469  0.92150909  4.373542e-02
#> 16      1   15 0.96524451      Gompertz  3.3417829992  0.95101930  1.422521e-02
#> 17      1   20 0.98938849   Exponential -0.0106682123  0.76571212  2.236764e-01
#> 18      1   20 0.98938849 Monomolecular  4.5458161350  0.98966621 -2.777227e-04
#> 19      1   20 0.98938849      Logistic  4.5351479227  0.98668405  2.704438e-03
#> 20      1   20 0.98938849      Gompertz  4.5404867709  0.98860235  7.861437e-04
#> 21      1   25 0.99703953   Exponential -0.0029648559  1.34654661 -3.495071e-01
#> 22      1   25 0.99703953 Monomolecular  5.8224088965  0.99712068 -8.114782e-05
#> 23      1   25 0.99703953      Logistic  5.8194440405  0.99786627 -8.267301e-04
#> 24      1   25 0.99703953      Gompertz  5.8209268348  0.99738693 -3.473912e-04
#> 25      1   30 0.99948686   Exponential -0.0005132751  2.36797578 -1.368489e+00
#> 26      1   30 0.99948686 Monomolecular  7.5749551934  0.99919773  2.891248e-04
#> 27      1   30 0.99948686      Logistic  7.5744419183  0.99966132 -1.744590e-04
#> 28      1   30 0.99948686      Gompertz  7.5746985668  0.99940295  8.390228e-05
#> 29      1    0 0.01146188   Exponential -4.4687282779  0.08006488 -6.860300e-02
#> 30      1    0 0.01146188 Monomolecular  0.0115280766 -0.71450693  7.259688e-01
#> 31      1    0 0.01146188      Logistic -4.4572002013  0.04461503 -3.315315e-02
#> 32      1    0 0.01146188      Gompertz -1.4971038666  0.01465304 -3.191155e-03
#> 33      1    5 0.38435986   Exponential -0.9561760334  0.14079847  2.435614e-01
#> 34      1    5 0.38435986 Monomolecular  0.4850926723  0.52228454 -1.379247e-01
#> 35      1    5 0.38435986      Logistic -0.4710833611  0.22763993  1.567199e-01
#> 36      1    5 0.38435986      Gompertz  0.0448132476  0.38138754  2.972315e-03
#> 37      1   10 0.78516524   Exponential -0.2418610805  0.24760180  5.375634e-01
#> 38      1   10 0.78516524 Monomolecular  1.5378861265  0.86689347 -8.172823e-02
#> 39      1   10 0.78516524      Logistic  1.2960250460  0.65037138  1.347939e-01
#> 40      1   10 0.78516524      Gompertz  1.4193917651  0.80250071 -1.733546e-02
#> 41      1   15 0.96445315   Exponential -0.0361940231  0.43542129  5.290319e-01
#> 42      1   15 0.96445315 Monomolecular  3.3369037088  0.96291234  1.540807e-03
#> 43      1   15 0.96445315      Logistic  3.3007096857  0.92150909  4.294406e-02
#> 44      1   15 0.96445315      Gompertz  3.3188612803  0.95101930  1.343385e-02
#> 45      1   20 0.98952391   Exponential -0.0105313496  0.76571212  2.238118e-01
#> 46      1   20 0.98952391 Monomolecular  4.5586598435  0.98966621 -1.423031e-04
#> 47      1   20 0.98952391      Logistic  4.5481284939  0.98668405  2.839858e-03
#> 48      1   20 0.98952391      Gompertz  4.5533987899  0.98860235  9.215633e-04
#> 49      1   25 0.99714065   Exponential -0.0028634491  1.34654661 -3.494060e-01
#> 50      1   25 0.99714065 Monomolecular  5.8571597687  0.99712068  1.996389e-05
#> 51      1   25 0.99714065      Logistic  5.8542963196  0.99786627 -7.256184e-04
#> 52      1   25 0.99714065      Gompertz  5.8557283858  0.99738693 -2.462795e-04
#> 53      1   30 0.99917978   Exponential -0.0008205569  2.36797578 -1.368796e+00
#> 54      1   30 0.99917978 Monomolecular  7.1059375918  0.99919773 -1.795210e-05
#> 55      1   30 0.99917978      Logistic  7.1051170349  0.99966132 -4.815359e-04
#> 56      1   30 0.99917978      Gompertz  7.1055273414  0.99940295 -2.231746e-04
#> 57      2    0 0.01222736   Exponential -4.4040789520  0.05667291 -4.444555e-02
#> 58      2    0 0.01222736 Monomolecular  0.0123027324 -0.58927829  6.015056e-01
#> 59      2    0 0.01222736      Logistic -4.3917762196  0.03443171 -2.220435e-02
#> 60      2    0 0.01222736      Gompertz -1.4825311460  0.01038509  1.842273e-03
#> 61      2    5 0.18241673   Exponential -1.7014614782  0.10487630  7.754043e-02
#> 62      2    5 0.18241673 Monomolecular  0.2014025234  0.27301583 -9.059910e-02
#> 63      2    5 0.18241673      Logistic -1.5000589548  0.12607439  5.634234e-02
#> 64      2    5 0.18241673      Gompertz -0.5314875748  0.18761908 -5.202351e-03
#> 65      2   10 0.53541286   Exponential -0.6247171298  0.19407929  3.413336e-01
#> 66      2   10 0.53541286 Monomolecular  0.7666061375  0.66745536 -1.320425e-01
#> 67      2   10 0.53541286      Logistic  0.1418890077  0.36853488  1.668780e-01
#> 68      2   10 0.53541286      Gompertz  0.4704563240  0.54169178 -6.278920e-03
#> 69      2   15 0.82586592   Exponential -0.1913228441  0.35915426  4.667117e-01
#> 70      2   15 0.82586592 Monomolecular  1.7479296971  0.84788398 -2.201806e-02
#> 71      2   15 0.82586592      Logistic  1.5566068530  0.70247435  1.233916e-01
#> 72      2   15 0.82586592      Gompertz  1.6537929947  0.79883199  2.703393e-02
#> 73      2   20 0.93132287   Exponential -0.0711492613  0.66463447  2.666884e-01
#> 74      2   20 0.93132287 Monomolecular  2.6783390510  0.93041751  9.053603e-04
#> 75      2   20 0.93132287      Logistic  2.6071897897  0.90522898  2.609389e-02
#> 76      2   20 0.93132287      Gompertz  2.6429753372  0.92100680  1.031607e-02
#> 77      2   25 0.97296399   Exponential -0.0274082105  1.22994217 -2.569782e-01
#> 78      2   25 0.97296399 Monomolecular  3.6105854617  0.96817086  4.793131e-03
#> 79      2   25 0.97296399      Logistic  3.5831772512  0.97477424 -1.810251e-03
#> 80      2   25 0.97296399      Gompertz  3.5969126567  0.97030230  2.661684e-03
#> 81      2   30 0.99071100   Exponential -0.0093324109  2.27607478 -1.285364e+00
#> 82      2   30 0.99071100 Monomolecular  4.6789244719  0.98544038  5.270618e-03
#> 83      2   30 0.99071100      Logistic  4.6695920610  0.99364385 -2.932848e-03
#> 84      2   30 0.99071100      Gompertz  4.6742618954  0.98901567  1.695333e-03
#> 85      2    0 0.01154863   Exponential -4.4611884746  0.05667291 -4.512428e-02
#> 86      2    0 0.01154863 Monomolecular  0.0116158332 -0.58927829  6.008269e-01
#> 87      2    0 0.01154863      Logistic -4.4495726413  0.03443171 -2.288308e-02
#> 88      2    0 0.01154863      Gompertz -1.4954152046  0.01038509  1.163539e-03
#> 89      2    5 0.18600157   Exponential -1.6820001380  0.10487630  8.112527e-02
#> 90      2    5 0.18600157 Monomolecular  0.2057968478  0.27301583 -8.701426e-02
#> 91      2    5 0.18600157      Logistic -1.4762032903  0.12607439  5.992719e-02
#> 92      2    5 0.18600157      Gompertz -0.5199836436  0.18761908 -1.617507e-03
#> 93      2   10 0.43727245   Exponential -0.8271988280  0.19407929  2.431932e-01
#> 94      2   10 0.43727245 Monomolecular  0.5749596893  0.66745536 -2.301829e-01
#> 95      2   10 0.43727245      Logistic -0.2522391387  0.36853488  6.873756e-02
#> 96      2   10 0.43727245      Gompertz  0.1897101921  0.54169178 -1.044193e-01
#> 97      2   15 0.81530917   Exponential -0.2041878886  0.35915426  4.561549e-01
#> 98      2   15 0.81530917 Monomolecular  1.6890720384  0.84788398 -3.257481e-02
#> 99      2   15 0.81530917      Logistic  1.4848841498  0.70247435  1.128348e-01
#> 100     2   15 0.81530917      Gompertz  1.5887146865  0.79883199  1.647718e-02
#> 101     2   20 0.91827755   Exponential -0.0852555908  0.66463447  2.536431e-01
#> 102     2   20 0.91827755 Monomolecular  2.5044265430  0.93041751 -1.213996e-02
#> 103     2   20 0.91827755      Logistic  2.4191709522  0.90522898  1.304857e-02
#> 104     2   20 0.91827755      Gompertz  2.4621015841  0.92100680 -2.729251e-03
#> 105     2   25 0.96852266   Exponential -0.0319833987  1.22994217 -2.614195e-01
#> 106     2   25 0.96852266 Monomolecular  3.4584873785  0.96817086  3.518049e-04
#> 107     2   25 0.96852266      Logistic  3.4265039798  0.97477424 -6.251576e-03
#> 108     2   25 0.96852266      Gompertz  3.4425383012  0.97030230 -1.779642e-03
#> 109     2   30 0.98534709   Exponential -0.0147613273  2.27607478 -1.290728e+00
#> 110     2   30 0.98534709 Monomolecular  4.2231161249  0.98544038 -9.329551e-05
#> 111     2   30 0.98534709      Logistic  4.2083547976  0.99364385 -8.296762e-03
#> 112     2   30 0.98534709      Gompertz  4.2157445403  0.98901567 -3.668581e-03
#> 
fit_multi(time_col = "time", intensity_col = "random_y", data = data)
#> $Parameters
#>     strata best_model         model         r       r_se   r_ci_lwr  r_ci_upr
#> 1 all_data          1      Logistic 0.3239974 0.02272449 0.27728651 0.3707082
#> 2 all_data          2      Gompertz 0.2481416 0.01765618 0.21184883 0.2844344
#> 3 all_data          3 Monomolecular 0.2059996 0.01796571 0.16907056 0.2429287
#> 4 all_data          4   Exponential 0.1179977 0.01901990 0.07890177 0.1570937
#>           v0     v0_se r_squared       RSE       CCC          y0     y0_ci_lwr
#> 1 -3.1988912 0.4096715 0.8866016 1.2024667 0.9398928  0.03920747  0.0172764590
#> 2 -1.4797559 0.3183014 0.8836780 0.9342775 0.9382474  0.01237752  0.0002141402
#> 3 -0.5012028 0.3238815 0.8348947 0.9506562 0.9100192 -0.65070550 -2.2121828056
#> 4 -2.6976884 0.3428862 0.5968272 1.0064387 0.7475164  0.06736104  0.0332899045
#>    y0_ci_upr
#> 1 0.08652683
#> 2 0.10198068
#> 3 0.15172056
#> 4 0.13630288
#> 
#> $Data
#>       strata time          y         model    linearized   predicted
#> 1   all_data    0 0.01000000   Exponential -4.6051701860  0.06736104
#> 2   all_data    0 0.01000000 Monomolecular  0.0100503359 -0.65070550
#> 3   all_data    0 0.01000000      Logistic -4.5951198501  0.03920747
#> 4   all_data    0 0.01000000      Gompertz -1.5271796258  0.01237752
#> 5   all_data    5 0.36633668   Exponential -1.0042024658  0.12151717
#> 6   all_data    5 0.36633668 Monomolecular  0.4562375135  0.41068550
#> 7   all_data    5 0.36633668      Logistic -0.5479649523  0.17095070
#> 8   all_data    5 0.36633668      Gompertz -0.0041936601  0.28082061
#> 9   all_data   10 0.76788928   Exponential -0.2641097288  0.21921309
#> 10  all_data   10 0.76788928 Monomolecular  1.4605407604  0.78961021
#> 11  all_data   10 0.76788928      Logistic  1.1964310316  0.51026914
#> 12  all_data   10 0.76788928      Gompertz  1.3313906230  0.69262322
#> 13  all_data   15 0.96524451   Exponential -0.0353738300  0.39545343
#> 14  all_data   15 0.96524451 Monomolecular  3.3594177769  0.92488924
#> 15  all_data   15 0.96524451      Logistic  3.3240439469  0.84038147
#> 16  all_data   15 0.96524451      Gompertz  3.3417829992  0.89923880
#> 17  all_data   20 0.98938849   Exponential -0.0106682123  0.71338536
#> 18  all_data   20 0.98938849 Monomolecular  4.5458161350  0.97318488
#> 19  all_data   20 0.98938849      Logistic  4.5351479227  0.96377317
#> 20  all_data   20 0.98938849      Gompertz  4.5404867709  0.96975408
#> 21  all_data   25 0.99703953   Exponential -0.0029648559  1.28692442
#> 22  all_data   25 0.99703953 Monomolecular  5.8224088965  0.99042680
#> 23  all_data   25 0.99703953      Logistic  5.8194440405  0.99261610
#> 24  all_data   25 0.99703953      Gompertz  5.8209268348  0.99115783
#> 25  all_data   30 0.99948686   Exponential -0.0005132751  2.32157058
#> 26  all_data   30 0.99948686 Monomolecular  7.5749551934  0.99658229
#> 27  all_data   30 0.99948686      Logistic  7.5744419183  0.99853001
#> 28  all_data   30 0.99948686      Gompertz  7.5746985668  0.99743495
#> 29  all_data    0 0.01146188   Exponential -4.4687282779  0.06736104
#> 30  all_data    0 0.01146188 Monomolecular  0.0115280766 -0.65070550
#> 31  all_data    0 0.01146188      Logistic -4.4572002013  0.03920747
#> 32  all_data    0 0.01146188      Gompertz -1.4971038666  0.01237752
#> 33  all_data    5 0.38435986   Exponential -0.9561760334  0.12151717
#> 34  all_data    5 0.38435986 Monomolecular  0.4850926723  0.41068550
#> 35  all_data    5 0.38435986      Logistic -0.4710833611  0.17095070
#> 36  all_data    5 0.38435986      Gompertz  0.0448132476  0.28082061
#> 37  all_data   10 0.78516524   Exponential -0.2418610805  0.21921309
#> 38  all_data   10 0.78516524 Monomolecular  1.5378861265  0.78961021
#> 39  all_data   10 0.78516524      Logistic  1.2960250460  0.51026914
#> 40  all_data   10 0.78516524      Gompertz  1.4193917651  0.69262322
#> 41  all_data   15 0.96445315   Exponential -0.0361940231  0.39545343
#> 42  all_data   15 0.96445315 Monomolecular  3.3369037088  0.92488924
#> 43  all_data   15 0.96445315      Logistic  3.3007096857  0.84038147
#> 44  all_data   15 0.96445315      Gompertz  3.3188612803  0.89923880
#> 45  all_data   20 0.98952391   Exponential -0.0105313496  0.71338536
#> 46  all_data   20 0.98952391 Monomolecular  4.5586598435  0.97318488
#> 47  all_data   20 0.98952391      Logistic  4.5481284939  0.96377317
#> 48  all_data   20 0.98952391      Gompertz  4.5533987899  0.96975408
#> 49  all_data   25 0.99714065   Exponential -0.0028634491  1.28692442
#> 50  all_data   25 0.99714065 Monomolecular  5.8571597687  0.99042680
#> 51  all_data   25 0.99714065      Logistic  5.8542963196  0.99261610
#> 52  all_data   25 0.99714065      Gompertz  5.8557283858  0.99115783
#> 53  all_data   30 0.99917978   Exponential -0.0008205569  2.32157058
#> 54  all_data   30 0.99917978 Monomolecular  7.1059375918  0.99658229
#> 55  all_data   30 0.99917978      Logistic  7.1051170349  0.99853001
#> 56  all_data   30 0.99917978      Gompertz  7.1055273414  0.99743495
#> 57  all_data    0 0.01222736   Exponential -4.4040789520  0.06736104
#> 58  all_data    0 0.01222736 Monomolecular  0.0123027324 -0.65070550
#> 59  all_data    0 0.01222736      Logistic -4.3917762196  0.03920747
#> 60  all_data    0 0.01222736      Gompertz -1.4825311460  0.01237752
#> 61  all_data    5 0.18241673   Exponential -1.7014614782  0.12151717
#> 62  all_data    5 0.18241673 Monomolecular  0.2014025234  0.41068550
#> 63  all_data    5 0.18241673      Logistic -1.5000589548  0.17095070
#> 64  all_data    5 0.18241673      Gompertz -0.5314875748  0.28082061
#> 65  all_data   10 0.53541286   Exponential -0.6247171298  0.21921309
#> 66  all_data   10 0.53541286 Monomolecular  0.7666061375  0.78961021
#> 67  all_data   10 0.53541286      Logistic  0.1418890077  0.51026914
#> 68  all_data   10 0.53541286      Gompertz  0.4704563240  0.69262322
#> 69  all_data   15 0.82586592   Exponential -0.1913228441  0.39545343
#> 70  all_data   15 0.82586592 Monomolecular  1.7479296971  0.92488924
#> 71  all_data   15 0.82586592      Logistic  1.5566068530  0.84038147
#> 72  all_data   15 0.82586592      Gompertz  1.6537929947  0.89923880
#> 73  all_data   20 0.93132287   Exponential -0.0711492613  0.71338536
#> 74  all_data   20 0.93132287 Monomolecular  2.6783390510  0.97318488
#> 75  all_data   20 0.93132287      Logistic  2.6071897897  0.96377317
#> 76  all_data   20 0.93132287      Gompertz  2.6429753372  0.96975408
#> 77  all_data   25 0.97296399   Exponential -0.0274082105  1.28692442
#> 78  all_data   25 0.97296399 Monomolecular  3.6105854617  0.99042680
#> 79  all_data   25 0.97296399      Logistic  3.5831772512  0.99261610
#> 80  all_data   25 0.97296399      Gompertz  3.5969126567  0.99115783
#> 81  all_data   30 0.99071100   Exponential -0.0093324109  2.32157058
#> 82  all_data   30 0.99071100 Monomolecular  4.6789244719  0.99658229
#> 83  all_data   30 0.99071100      Logistic  4.6695920610  0.99853001
#> 84  all_data   30 0.99071100      Gompertz  4.6742618954  0.99743495
#> 85  all_data    0 0.01154863   Exponential -4.4611884746  0.06736104
#> 86  all_data    0 0.01154863 Monomolecular  0.0116158332 -0.65070550
#> 87  all_data    0 0.01154863      Logistic -4.4495726413  0.03920747
#> 88  all_data    0 0.01154863      Gompertz -1.4954152046  0.01237752
#> 89  all_data    5 0.18600157   Exponential -1.6820001380  0.12151717
#> 90  all_data    5 0.18600157 Monomolecular  0.2057968478  0.41068550
#> 91  all_data    5 0.18600157      Logistic -1.4762032903  0.17095070
#> 92  all_data    5 0.18600157      Gompertz -0.5199836436  0.28082061
#> 93  all_data   10 0.43727245   Exponential -0.8271988280  0.21921309
#> 94  all_data   10 0.43727245 Monomolecular  0.5749596893  0.78961021
#> 95  all_data   10 0.43727245      Logistic -0.2522391387  0.51026914
#> 96  all_data   10 0.43727245      Gompertz  0.1897101921  0.69262322
#> 97  all_data   15 0.81530917   Exponential -0.2041878886  0.39545343
#> 98  all_data   15 0.81530917 Monomolecular  1.6890720384  0.92488924
#> 99  all_data   15 0.81530917      Logistic  1.4848841498  0.84038147
#> 100 all_data   15 0.81530917      Gompertz  1.5887146865  0.89923880
#> 101 all_data   20 0.91827755   Exponential -0.0852555908  0.71338536
#> 102 all_data   20 0.91827755 Monomolecular  2.5044265430  0.97318488
#> 103 all_data   20 0.91827755      Logistic  2.4191709522  0.96377317
#> 104 all_data   20 0.91827755      Gompertz  2.4621015841  0.96975408
#> 105 all_data   25 0.96852266   Exponential -0.0319833987  1.28692442
#> 106 all_data   25 0.96852266 Monomolecular  3.4584873785  0.99042680
#> 107 all_data   25 0.96852266      Logistic  3.4265039798  0.99261610
#> 108 all_data   25 0.96852266      Gompertz  3.4425383012  0.99115783
#> 109 all_data   30 0.98534709   Exponential -0.0147613273  2.32157058
#> 110 all_data   30 0.98534709 Monomolecular  4.2231161249  0.99658229
#> 111 all_data   30 0.98534709      Logistic  4.2083547976  0.99853001
#> 112 all_data   30 0.98534709      Gompertz  4.2157445403  0.99743495
#>          residual
#> 1   -0.0573610415
#> 2    0.6607054954
#> 3   -0.0292074696
#> 4   -0.0023775159
#> 5    0.2448195155
#> 6   -0.0443488167
#> 7    0.1953859876
#> 8    0.0855160714
#> 9    0.5486761812
#> 10  -0.0217209368
#> 11   0.2576201327
#> 12   0.0752660527
#> 13   0.5697910858
#> 14   0.0403552763
#> 15   0.1248630425
#> 16   0.0660057073
#> 17   0.2760031326
#> 18   0.0162036093
#> 19   0.0256153229
#> 20   0.0196344153
#> 21  -0.2898848836
#> 22   0.0066127373
#> 23   0.0044234300
#> 24   0.0058817098
#> 25  -1.3220837208
#> 26   0.0029045631
#> 27   0.0009568453
#> 28   0.0020519073
#> 29  -0.0558991585
#> 30   0.6621673783
#> 31  -0.0277455867
#> 32  -0.0009156330
#> 33   0.2628426897
#> 34  -0.0263256424
#> 35   0.2134091619
#> 36   0.1035392456
#> 37   0.5659521504
#> 38  -0.0044449676
#> 39   0.2748961019
#> 40   0.0925420219
#> 41   0.5689997234
#> 42   0.0395639139
#> 43   0.1240716801
#> 44   0.0652143449
#> 45   0.2761385522
#> 46   0.0163390289
#> 47   0.0257507425
#> 48   0.0197698350
#> 49  -0.2897837719
#> 50   0.0067138490
#> 51   0.0045245417
#> 52   0.0059828215
#> 53  -1.3223907977
#> 54   0.0025974863
#> 55   0.0006497684
#> 56   0.0017448305
#> 57  -0.0551336782
#> 58   0.6629328586
#> 59  -0.0269801064
#> 60  -0.0001501527
#> 61   0.0608995622
#> 62  -0.2282687700
#> 63   0.0114660343
#> 64  -0.0984038819
#> 65   0.3161997651
#> 66  -0.2541973529
#> 67   0.0251437166
#> 68  -0.1572103634
#> 69   0.4304124935
#> 70  -0.0990233160
#> 71  -0.0145155498
#> 72  -0.0733728850
#> 73   0.2179375126
#> 74  -0.0418620106
#> 75  -0.0324502970
#> 76  -0.0384312046
#> 77  -0.3139604322
#> 78  -0.0174628112
#> 79  -0.0196521186
#> 80  -0.0181938388
#> 81  -1.3308595765
#> 82  -0.0058712925
#> 83  -0.0078190104
#> 84  -0.0067239483
#> 85  -0.0558124116
#> 86   0.6622541253
#> 87  -0.0276588397
#> 88  -0.0008288860
#> 89   0.0644844060
#> 90  -0.2246839262
#> 91   0.0150508781
#> 92  -0.0948190381
#> 93   0.2180593534
#> 94  -0.3523377646
#> 95  -0.0729966951
#> 96  -0.2553507751
#> 97   0.4198557438
#> 98  -0.1095800657
#> 99  -0.0250722995
#> 100 -0.0839296347
#> 101  0.2048921925
#> 102 -0.0549073308
#> 103 -0.0454956172
#> 104 -0.0514765248
#> 105 -0.3184017579
#> 106 -0.0219041369
#> 107 -0.0240934442
#> 108 -0.0226351644
#> 109 -1.3362234904
#> 110 -0.0112352064
#> 111 -0.0131829243
#> 112 -0.0120878622
#>