Estimate model parameters for multiple disease progress curves

Estimate model parameters for multiple disease progress curves

fit_multi(time_col,
             intensity_col,
             data,
             strata_cols ,
             starting_par = list(y0 = 0.01, r = 0.03, K =  0.8),
             maxiter=500,
             nlin = FALSE,
             estimate_K = FALSE)

Arguments

time_col
Character name specifying the column for the time. eg: time_col = "days".

intensity_col
Character name specifying the column for the disease intensity.

data
data.frame object containing the variables for model fitting.

strata_cols
Character name or vector specifying the columns for stratification.

starting_par
Starting value for initial inoculun (y0) and apparent infection rate (r). Please informe in that especific order

maxiter
Maximum number of iterations. Only used if is nlin = TRUE

nlin
Logical. If FALSE estimates parameters using data linearization. If nlin=TRUE, estimates nonlinear approach. fit_nlin.

estimate_K
Logical. If nlin=TRUE, estimates maximum disease intensity. (default: nlin=FALSE) fit_nlin2.

Value

Returns a data.frame containing estimated parameters for individual strata levels.

See also

Examples

set.seed(1)
# create stratified dataset
data_A1 = sim_gompertz(N = 30, y0 = 0.01,dt = 5, r = 0.3, alpha = 0.5, n = 4)
data_A1 = dplyr::mutate(data_A1,
                        fun = "A",
                        cultivar = "BR1")
set.seed(1)
data_B1 = sim_gompertz(N = 30, y0 = 0.01, dt = 5, r = 0.2, alpha = 0.5, n = 4)
data_B1 = dplyr::mutate(data_B1,
                        fun = "B",
                        cultivar = "BR1")
set.seed(1)
data_A2 = sim_gompertz(N = 30, y0 = 0.01,dt = 5, r = 0.1, alpha = 0.5, n = 4)
data_A2 = dplyr::mutate(data_A2,
                        fun = "A",
                        cultivar = "BR2")
set.seed(1)
data_B2 = sim_gompertz(N = 30, y0 = 0.01,dt = 5, r = 0.1, alpha = 0.5, n = 4)
data_B2 = dplyr::mutate(data_B2,
                        fun = "B",
                        cultivar = "BR2")

data = dplyr::bind_rows(data_A1, data_B1,data_A2, data_B2)

fit_multi(time_col = "time",
             intensity_col = "random_y",
             data = data,
             strata_col = c("fun","cultivar"),
             starting_par = list(y0 = 0.01, r = 0.03),
             maxiter = 1024,
             nlin = FALSE,
             estimate_K = FALSE)

#> $Parameters
#>    fun cultivar best_model         model          r        r_se   r_ci_lwr
#> 1    A      BR1          1      Gompertz 0.29808279 0.009443015 0.27867239
#> 2    A      BR1          2 Monomolecular 0.25898305 0.010208285 0.23799962
#> 3    A      BR1          3      Logistic 0.36784872 0.017501377 0.33187413
#> 4    A      BR1          4   Exponential 0.10886568 0.019616106 0.06854419
#> 5    B      BR1          1      Gompertz 0.19808503 0.008164125 0.18130343
#> 6    B      BR1          2 Monomolecular 0.15418807 0.008001841 0.13774005
#> 7    B      BR1          3      Logistic 0.27499434 0.015501019 0.24313154
#> 8    B      BR1          4   Exponential 0.12080627 0.016495660 0.08689896
#> 9    A      BR2          1      Gompertz 0.09684563 0.006070602 0.08436733
#> 10   A      BR2          2      Logistic 0.18264336 0.012656741 0.15662706
#> 11   A      BR2          3 Monomolecular 0.05130200 0.004408226 0.04224076
#> 12   A      BR2          4   Exponential 0.13134136 0.011842997 0.10699773
#> 13   B      BR2          1      Gompertz 0.09684563 0.006070602 0.08436733
#> 14   B      BR2          2      Logistic 0.18264336 0.012656741 0.15662706
#> 15   B      BR2          3 Monomolecular 0.05130200 0.004408226 0.04224076
#> 16   B      BR2          4   Exponential 0.13134136 0.011842997 0.10699773
#>      r_ci_upr         v0      v0_se r_squared       RSE       CCC          y0
#> 1  0.31749318 -1.3853100 0.17023637 0.9745707 0.4996774 0.9871216  0.01838786
#> 2  0.27996648 -0.4993432 0.18403248 0.9611727 0.5401717 0.9802020 -0.64763876
#> 3  0.40382332 -2.9402875 0.31551056 0.9444169 0.9260858 0.9714140  0.05019756
#> 4  0.14918716 -2.4409443 0.35363439 0.5422567 1.0379868 0.7031991  0.08707858
#> 5  0.21486663 -1.4159626 0.14718085 0.9577020 0.4320049 0.9783941  0.01623719
#> 6  0.17063609 -0.3728576 0.14425524 0.9345578 0.4234176 0.9661720 -0.45187757
#> 7  0.30685714 -3.1909969 0.27944859 0.9236914 0.8202368 0.9603322  0.03950593
#> 8  0.15471359 -2.8181393 0.29737974 0.6735060 0.8728683 0.8049042  0.05971695
#> 9  0.10932393 -1.4598600 0.10943933 0.9073101 0.3212260 0.9514028  0.01349607
#> 10 0.20865966 -3.7292757 0.22817264 0.8890027 0.6697318 0.9412403  0.02344725
#> 11 0.06036324 -0.1486216 0.07947043 0.8389476 0.2332614 0.9124214 -0.16023391
#> 12 0.15568499 -3.5806541 0.21350266 0.8254950 0.6266725 0.9044067  0.02785747
#> 13 0.10932393 -1.4598600 0.10943933 0.9073101 0.3212260 0.9514028  0.01349607
#> 14 0.20865966 -3.7292757 0.22817264 0.8890027 0.6697318 0.9412403  0.02344725
#> 15 0.06036324 -0.1486216 0.07947043 0.8389476 0.2332614 0.9124214 -0.16023391
#> 16 0.15568499 -3.5806541 0.21350266 0.8254950 0.6266725 0.9044067  0.02785747
#>       y0_ci_lwr   y0_ci_upr
#> 1   0.003446951  0.05983360
#> 2  -1.405186351 -0.12869154
#> 3   0.026887747  0.09180884
#> 4   0.042093993  0.18013685
#> 5   0.003787148  0.04760688
#> 6  -0.953023103 -0.07932593
#> 7   0.022633917  0.06807882
#> 8   0.032405985  0.11004493
#> 9   0.004555328  0.03212729
#> 10  0.014798924  0.03695995
#> 11 -0.366121131  0.01462418
#> 12  0.017961673  0.04320526
#> 13  0.004555328  0.03212729
#> 14  0.014798924  0.03695995
#> 15 -0.366121131  0.01462418
#> 16  0.017961673  0.04320526
#> 
#> $Data
#>     fun cultivar time          y         model    linearized   predicted
#> 1     A      BR1    0 0.01000000   Exponential -4.6051701860  0.08707858
#> 2     A      BR1    0 0.01000000 Monomolecular  0.0100503359 -0.64763876
#> 3     A      BR1    0 0.01000000      Logistic -4.5951198501  0.05019756
#> 4     A      BR1    0 0.01000000      Gompertz -1.5271796258  0.01838786
#> 5     A      BR1    5 0.37899741   Exponential -0.9702258977  0.15007562
#> 6     A      BR1    5 0.37899741 Monomolecular  0.4764200324  0.54867701
#> 7     A      BR1    5 0.37899741      Logistic -0.4938058653  0.24954435
#> 8     A      BR1    5 0.37899741      Gompertz  0.0302263503  0.40647578
#> 9     A      BR1   10 0.72705035   Exponential -0.3187595472  0.25864791
#> 10    A      BR1   10 0.72705035 Monomolecular  1.2984679314  0.87637312
#> 11    A      BR1   10 0.72705035      Logistic  0.9797083842  0.67660205
#> 12    A      BR1   10 0.72705035      Gompertz  1.1433182309  0.81643868
#> 13    A      BR1   15 0.98792327   Exponential -0.0121502482  0.44576687
#> 14    A      BR1   15 0.98792327 Monomolecular  4.4164746553  0.96613599
#> 15    A      BR1   15 0.98792327      Logistic  4.4043244071  0.92939569
#> 16    A      BR1   15 0.98792327      Gompertz  4.4104056824  0.95534052
#> 17    A      BR1   20 0.99049783   Exponential -0.0095475994  0.76825713
#> 18    A      BR1   20 0.99049783 Monomolecular  4.6562355322  0.99072393
#> 19    A      BR1   20 0.99049783      Logistic  4.6466879328  0.98806988
#> 20    A      BR1   20 0.99049783      Gompertz  4.6514655307  0.98976035
#> 21    A      BR1   25 0.99641493   Exponential -0.0035915106  1.32405313
#> 22    A      BR1   25 0.99641493 Monomolecular  5.6309775912  0.99745909
#> 23    A      BR1   25 0.99641493      Logistic  5.6273860805  0.99808464
#> 24    A      BR1   25 0.99641493      Gompertz  5.6291823733  0.99768401
#> 25    A      BR1   30 0.99956995   Exponential -0.0004301469  2.28194003
#> 26    A      BR1   30 0.99956995 Monomolecular  7.7515989041  0.99930399
#> 27    A      BR1   30 0.99956995      Logistic  7.7511687572  0.99969509
#> 28    A      BR1   30 0.99956995      Gompertz  7.7513838384  0.99947779
#> 29    A      BR1    0 0.01365471   Exponential -4.2936709604  0.08707858
#> 30    A      BR1    0 0.01365471 Monomolecular  0.0137487902 -0.64763876
#> 31    A      BR1    0 0.01365471      Logistic -4.2799221701  0.05019756
#> 32    A      BR1    0 0.01365471      Gompertz -1.4571420688  0.01838786
#> 33    A      BR1    5 0.42405535   Exponential -0.8578912912  0.15007562
#> 34    A      BR1    5 0.42405535 Monomolecular  0.5517437157  0.54867701
#> 35    A      BR1    5 0.42405535      Logistic -0.3061475755  0.24954435
#> 36    A      BR1    5 0.42405535      Gompertz  0.1532778878  0.40647578
#> 37    A      BR1   10 0.77024027   Exponential -0.2610527704  0.25864791
#> 38    A      BR1   10 0.77024027 Monomolecular  1.4707211797  0.87637312
#> 39    A      BR1   10 0.77024027      Logistic  1.2096684092  0.67660205
#> 40    A      BR1   10 0.77024027      Gompertz  1.3430327065  0.81643868
#> 41    A      BR1   15 0.98594486   Exponential -0.0141548468  0.44576687
#> 42    A      BR1   15 0.98594486 Monomolecular  4.2647672585  0.96613599
#> 43    A      BR1   15 0.98594486      Logistic  4.2506124117  0.92939569
#> 44    A      BR1   15 0.98594486      Gompertz  4.2576981834  0.95534052
#> 45    A      BR1   20 0.99083638   Exponential -0.0092058609  0.76825713
#> 46    A      BR1   20 0.99083638 Monomolecular  4.6925143412  0.99072393
#> 47    A      BR1   20 0.99083638      Logistic  4.6833084803  0.98806988
#> 48    A      BR1   20 0.99083638      Gompertz  4.6879149419  0.98976035
#> 49    A      BR1   25 0.99666771   Exponential -0.0033378541  1.32405313
#> 50    A      BR1   25 0.99666771 Monomolecular  5.7040956408  0.99745909
#> 51    A      BR1   25 0.99666771      Logistic  5.7007577867  0.99808464
#> 52    A      BR1   25 0.99666771      Gompertz  5.7024271780  0.99768401
#> 53    A      BR1   30 0.99880225   Exponential -0.0011984645  2.28194003
#> 54    A      BR1   30 0.99880225 Monomolecular  6.7273133223  0.99930399
#> 55    A      BR1   30 0.99880225      Logistic  6.7261148578  0.99969509
#> 56    A      BR1   30 0.99880225      Gompertz  6.7267141499  0.99947779
#> 57    A      BR1    0 0.01556841   Exponential -4.1625115434  0.08707858
#> 58    A      BR1    0 0.01556841 Monomolecular  0.0156908684 -0.64763876
#> 59    A      BR1    0 0.01556841      Logistic -4.1468206750  0.05019756
#> 60    A      BR1    0 0.01556841      Gompertz -1.4261186285  0.01838786
#> 61    A      BR1    5 0.35273318   Exponential -1.0420433714  0.15007562
#> 62    A      BR1    5 0.35273318 Monomolecular  0.4349966741  0.54867701
#> 63    A      BR1    5 0.35273318      Logistic -0.6070466973  0.24954435
#> 64    A      BR1    5 0.35273318      Gompertz -0.0411835657  0.40647578
#> 65    A      BR1   10 0.79379647   Exponential -0.2309281825  0.25864791
#> 66    A      BR1   10 0.79379647 Monomolecular  1.5788916004  0.87637312
#> 67    A      BR1   10 0.79379647      Logistic  1.3479634179  0.67660205
#> 68    A      BR1   10 0.79379647      Gompertz  1.4656485153  0.81643868
#> 69    A      BR1   15 0.97248821   Exponential -0.0278973294  0.44576687
#> 70    A      BR1   15 0.97248821 Monomolecular  3.5931405541  0.96613599
#> 71    A      BR1   15 0.97248821      Logistic  3.5652432248  0.92939569
#> 72    A      BR1   15 0.97248821      Gompertz  3.5792243168  0.95534052
#> 73    A      BR1   20 0.99325689   Exponential -0.0067659441  0.76825713
#> 74    A      BR1   20 0.99325689 Monomolecular  4.9992345290  0.99072393
#> 75    A      BR1   20 0.99325689      Logistic  4.9924685848  0.98806988
#> 76    A      BR1   20 0.99325689      Gompertz  4.9958534643  0.98976035
#> 77    A      BR1   25 0.99820948   Exponential -0.0017921284  1.32405313
#> 78    A      BR1   25 0.99820948 Monomolecular  6.3252472723  0.99745909
#> 79    A      BR1   25 0.99820948      Logistic  6.3234551439  0.99808464
#> 80    A      BR1   25 0.99820948      Gompertz  6.3243513419  0.99768401
#> 81    A      BR1   30 0.99969255   Exponential -0.0003074960  2.28194003
#> 82    A      BR1   30 0.99969255 Monomolecular  8.0872022495  0.99930399
#> 83    A      BR1   30 0.99969255      Logistic  8.0868947535  0.99969509
#> 84    A      BR1   30 0.99969255      Gompertz  8.0870485054  0.99947779
#> 85    A      BR1    0 0.01387157   Exponential -4.2779135191  0.08707858
#> 86    A      BR1    0 0.01387157 Monomolecular  0.0139686841 -0.64763876
#> 87    A      BR1    0 0.01387157      Logistic -4.2639448350  0.05019756
#> 88    A      BR1    0 0.01387157      Gompertz -1.4534653952  0.01838786
#> 89    A      BR1    5 0.36646396   Exponential -1.0038551052  0.15007562
#> 90    A      BR1    5 0.36646396 Monomolecular  0.4564383864  0.54867701
#> 91    A      BR1    5 0.36646396      Logistic -0.5474167188  0.24954435
#> 92    A      BR1    5 0.36646396      Gompertz -0.0038476934  0.40647578
#> 93    A      BR1   10 0.63307556   Exponential -0.4571654888  0.25864791
#> 94    A      BR1   10 0.63307556 Monomolecular  1.0025993499  0.87637312
#> 95    A      BR1   10 0.63307556      Logistic  0.5454338611  0.67660205
#> 96    A      BR1   10 0.63307556      Gompertz  0.7827098337  0.81643868
#> 97    A      BR1   15 0.96481124   Exponential -0.0358228062  0.44576687
#> 98    A      BR1   15 0.96481124 Monomolecular  3.3470284789  0.96613599
#> 99    A      BR1   15 0.96481124      Logistic  3.3112056728  0.92939569
#> 100   A      BR1   15 0.96481124      Gompertz  3.3291705450  0.95534052
#> 101   A      BR1   20 0.98833398   Exponential -0.0117345973  0.76825713
#> 102   A      BR1   20 0.98833398 Monomolecular  4.4510753304  0.99072393
#> 103   A      BR1   20 0.98833398      Logistic  4.4393407331  0.98806988
#> 104   A      BR1   20 0.98833398      Gompertz  4.4452137693  0.98976035
#> 105   A      BR1   25 0.99725826   Exponential -0.0027455006  1.32405313
#> 106   A      BR1   25 0.99725826 Monomolecular  5.8991642912  0.99745909
#> 107   A      BR1   25 0.99725826      Logistic  5.8964187906  0.99808464
#> 108   A      BR1   25 0.99725826      Gompertz  5.8977918550  0.99768401
#> 109   A      BR1   30 0.99901361   Exponential -0.0009868731  2.28194003
#> 110   A      BR1   30 0.99901361 Monomolecular  6.9214624704  0.99930399
#> 111   A      BR1   30 0.99901361      Logistic  6.9204755973  0.99969509
#> 112   A      BR1   30 0.99901361      Gompertz  6.9209690744  0.99947779
#> 113   B      BR1    0 0.01000000   Exponential -4.6051701860  0.05971695
#> 114   B      BR1    0 0.01000000 Monomolecular  0.0100503359 -0.45187757
#> 115   B      BR1    0 0.01000000      Logistic -4.5951198501  0.03950593
#> 116   B      BR1    0 0.01000000      Gompertz -1.5271796258  0.01623719
#> 117   B      BR1    5 0.19753751   Exponential -1.6218267983  0.10925093
#> 118   B      BR1    5 0.19753751 Monomolecular  0.2200701641  0.32839355
#> 119   B      BR1    5 0.19753751      Logistic -1.4017566342  0.13991150
#> 120   B      BR1    5 0.19753751      Gompertz -0.4835531672  0.21644632
#> 121   B      BR1   10 0.43231262   Exponential -0.8386063008  0.19987230
#> 122   B      BR1   10 0.43231262 Monomolecular  0.5661843947  0.68932971
#> 123   B      BR1   10 0.43231262      Logistic -0.2724219061  0.39148996
#> 124   B      BR1   10 0.43231262      Gompertz  0.1760139308  0.56641743
#> 125   B      BR1   15 0.92505595   Exponential -0.0779010573  0.36566223
#> 126   B      BR1   15 0.92505595 Monomolecular  2.5910134382  0.85629079
#> 127   B      BR1   15 0.92505595      Logistic  2.5131123809  0.71786952
#> 128   B      BR1   15 0.92505595      Gompertz  2.5523157541  0.80967421
#> 129   B      BR1   20 0.93136076   Exponential -0.0711085783  0.66897146
#> 130   B      BR1   20 0.93136076 Monomolecular  2.6788909111  0.93352330
#> 131   B      BR1   20 0.93136076      Logistic  2.6077823327  0.90961062
#> 132   B      BR1   20 0.93136076      Gompertz  2.6435472976  0.92458050
#> 133   B      BR1   25 0.95729415   Exponential -0.0436445691  1.22386941
#> 134   B      BR1   25 0.95729415 Monomolecular  3.1534193416  0.96924935
#> 135   B      BR1   25 0.95729415      Logistic  3.1097747725  0.97549007
#> 136   B      BR1   25 0.95729415      Gompertz  3.1316764245  0.97129518
#> 137   B      BR1   30 0.99138346   Exponential -0.0086538738  2.23904370
#> 138   B      BR1   30 0.99138346 Monomolecular  4.7540720391  0.98577543
#> 139   B      BR1   30 0.99138346      Logistic  4.7454181653  0.99368714
#> 140   B      BR1   30 0.99138346      Gompertz  4.7497482226  0.98924078
#> 141   B      BR1    0 0.01365471   Exponential -4.2936709604  0.05971695
#> 142   B      BR1    0 0.01365471 Monomolecular  0.0137487902 -0.45187757
#> 143   B      BR1    0 0.01365471      Logistic -4.2799221701  0.03950593
#> 144   B      BR1    0 0.01365471      Gompertz -1.4571420688  0.01623719
#> 145   B      BR1    5 0.22694693   Exponential -1.4830390974  0.10925093
#> 146   B      BR1    5 0.22694693 Monomolecular  0.2574075723  0.32839355
#> 147   B      BR1    5 0.22694693      Logistic -1.2256315251  0.13991150
#> 148   B      BR1    5 0.22694693      Gompertz -0.3940934265  0.21644632
#> 149   B      BR1   10 0.49824487   Exponential -0.6966636086  0.19987230
#> 150   B      BR1   10 0.49824487 Monomolecular  0.6896430745  0.68932971
#> 151   B      BR1   10 0.49824487      Logistic -0.0070205341  0.39148996
#> 152   B      BR1   10 0.49824487      Gompertz  0.3614526123  0.56641743
#> 153   B      BR1   15 0.91825459   Exponential -0.0852805962  0.36566223
#> 154   B      BR1   15 0.91825459 Monomolecular  2.5041456113  0.85629079
#> 155   B      BR1   15 0.91825459      Logistic  2.4188650151  0.71786952
#> 156   B      BR1   15 0.91825459      Gompertz  2.4618083274  0.80967421
#> 157   B      BR1   20 0.93360358   Exponential -0.0687033615  0.66897146
#> 158   B      BR1   20 0.93360358 Monomolecular  2.7121121663  0.93352330
#> 159   B      BR1   20 0.93360358      Logistic  2.6434088048  0.90961062
#> 160   B      BR1   20 0.93360358      Gompertz  2.6779571508  0.92458050
#> 161   B      BR1   25 0.96024479   Exponential -0.0405670337  1.22386941
#> 162   B      BR1   25 0.96024479 Monomolecular  3.2250144665  0.96924935
#> 163   B      BR1   25 0.96024479      Logistic  3.1844474328  0.97549007
#> 164   B      BR1   25 0.96024479      Gompertz  3.2047995189  0.97129518
#> 165   B      BR1   30 0.97622067   Exponential -0.0240666172  2.23904370
#> 166   B      BR1   30 0.97622067 Monomolecular  3.7389387521  0.98577543
#> 167   B      BR1   30 0.97622067      Logistic  3.7148721349  0.99368714
#> 168   B      BR1   30 0.97622067      Gompertz  3.7269295768  0.98924078
#> 169   B      BR1    0 0.01556841   Exponential -4.1625115434  0.05971695
#> 170   B      BR1    0 0.01556841 Monomolecular  0.0156908684 -0.45187757
#> 171   B      BR1    0 0.01556841      Logistic -4.1468206750  0.03950593
#> 172   B      BR1    0 0.01556841      Gompertz -1.4261186285  0.01623719
#> 173   B      BR1    5 0.18039478   Exponential -1.7126075912  0.10925093
#> 174   B      BR1    5 0.18039478 Monomolecular  0.1989324973  0.32839355
#> 175   B      BR1    5 0.18039478      Logistic -1.5136750940  0.13991150
#> 176   B      BR1    5 0.18039478      Gompertz -0.5380171162  0.21644632
#> 177   B      BR1   10 0.53420496   Exponential -0.6269756916  0.19987230
#> 178   B      BR1   10 0.53420496 Monomolecular  0.7640095721  0.68932971
#> 179   B      BR1   10 0.53420496      Logistic  0.1370338805  0.39148996
#> 180   B      BR1   10 0.53420496      Gompertz  0.4668475085  0.56641743
#> 181   B      BR1   15 0.87199333   Exponential -0.1369735034  0.36566223
#> 182   B      BR1   15 0.87199333 Monomolecular  2.0556729121  0.85629079
#> 183   B      BR1   15 0.87199333      Logistic  1.9186994086  0.71786952
#> 184   B      BR1   15 0.87199333      Gompertz  1.9879677773  0.80967421
#> 185   B      BR1   20 0.94963898   Exponential -0.0516733836  0.66897146
#> 186   B      BR1   20 0.94963898 Monomolecular  2.9885378925  0.93352330
#> 187   B      BR1   20 0.94963898      Logistic  2.9368645089  0.90961062
#> 188   B      BR1   20 0.94963898      Gompertz  2.9628124540  0.92458050
#> 189   B      BR1   25 0.97824154   Exponential -0.0219986681  1.22386941
#> 190   B      BR1   25 0.97824154 Monomolecular  3.8277525396  0.96924935
#> 191   B      BR1   25 0.97824154      Logistic  3.8057538715  0.97549007
#> 192   B      BR1   25 0.97824154      Gompertz  3.8167733697  0.97129518
#> 193   B      BR1   30 0.99380506   Exponential -0.0062142048  2.23904370
#> 194   B      BR1   30 0.99380506 Monomolecular  5.0840229990  0.98577543
#> 195   B      BR1   30 0.99380506      Logistic  5.0778087942  0.99368714
#> 196   B      BR1   30 0.99380506      Gompertz  5.0809175056  0.98924078
#> 197   B      BR1    0 0.01387157   Exponential -4.2779135191  0.05971695
#> 198   B      BR1    0 0.01387157 Monomolecular  0.0139686841 -0.45187757
#> 199   B      BR1    0 0.01387157      Logistic -4.2639448350  0.03950593
#> 200   B      BR1    0 0.01387157      Gompertz -1.4534653952  0.01623719
#> 201   B      BR1    5 0.18935689   Exponential -1.6641217243  0.10925093
#> 202   B      BR1    5 0.18935689 Monomolecular  0.2099273865  0.32839355
#> 203   B      BR1    5 0.18935689      Logistic -1.4541943378  0.13991150
#> 204   B      BR1    5 0.18935689      Gompertz -0.5092974914  0.21644632
#> 205   B      BR1   10 0.28885393   Exponential -1.2418341453  0.19987230
#> 206   B      BR1   10 0.28885393 Monomolecular  0.3408774296  0.68932971
#> 207   B      BR1   10 0.28885393      Logistic -0.9009567157  0.39148996
#> 208   B      BR1   10 0.28885393      Gompertz -0.2165894362  0.56641743
#> 209   B      BR1   15 0.84560146   Exponential -0.1677071221  0.36566223
#> 210   B      BR1   15 0.84560146 Monomolecular  1.8682180742  0.85629079
#> 211   B      BR1   15 0.84560146      Logistic  1.7005109521  0.71786952
#> 212   B      BR1   15 0.84560146      Gompertz  1.7855361419  0.80967421
#> 213   B      BR1   20 0.91702568   Exponential -0.0866197990  0.66897146
#> 214   B      BR1   20 0.91702568 Monomolecular  2.4892241578  0.93352330
#> 215   B      BR1   20 0.91702568      Logistic  2.4026043587  0.90961062
#> 216   B      BR1   20 0.91702568      Gompertz  2.4462268633  0.92458050
#> 217   B      BR1   25 0.96713822   Exponential -0.0334138528  1.22386941
#> 218   B      BR1   25 0.96713822 Monomolecular  3.4154451169  0.96924935
#> 219   B      BR1   25 0.96713822      Logistic  3.3820312641  0.97549007
#> 220   B      BR1   25 0.96713822      Gompertz  3.3987847103  0.97129518
#> 221   B      BR1   30 0.98039528   Exponential -0.0197994430  2.23904370
#> 222   B      BR1   30 0.98039528 Monomolecular  3.9319848586  0.98577543
#> 223   B      BR1   30 0.98039528      Logistic  3.9121854156  0.99368714
#> 224   B      BR1   30 0.98039528      Gompertz  3.9221014711  0.98924078
#> 225   A      BR2    0 0.01000000   Exponential -4.6051701860  0.02785747
#> 226   A      BR2    0 0.01000000 Monomolecular  0.0100503359 -0.16023391
#> 227   A      BR2    0 0.01000000      Logistic -4.5951198501  0.02344725
#> 228   A      BR2    0 0.01000000      Gompertz -1.5271796258  0.01349607
#> 229   A      BR2    5 0.06650718   Exponential -2.7104454417  0.05372122
#> 230   A      BR2    5 0.06650718 Monomolecular  0.0688220022  0.10227218
#> 231   A      BR2    5 0.06650718      Logistic -2.6416234395  0.05646253
#> 232   A      BR2    5 0.06650718      Gompertz -0.9971129910  0.07045064
#> 233   A      BR2   10 0.12108868   Exponential -2.1112320932  0.10359767
#> 234   A      BR2   10 0.12108868 Monomolecular  0.1290712760  0.30538555
#> 235   A      BR2   10 0.12108868      Logistic -1.9821608172  0.12978714
#> 236   A      BR2   10 0.12108868      Gompertz -0.7472717075  0.19502858
#> 237   A      BR2   15 0.54118615   Exponential -0.6139919688  0.19978098
#> 238   A      BR2   15 0.54118615 Monomolecular  0.7791107132  0.46254397
#> 239   A      BR2   15 0.54118615      Logistic  0.1651187445  0.27098609
#> 240   A      BR2   15 0.54118615      Gompertz  0.4877734311  0.36523971
#> 241   A      BR2   20 0.57718069   Exponential -0.5495999042  0.38526387
#> 242   A      BR2   20 0.57718069 Monomolecular  0.8608103584  0.58414486
#> 243   A      BR2   20 0.57718069      Logistic  0.3112104542  0.48090716
#> 244   A      BR2   20 0.57718069      Gompertz  0.5985647123  0.53761644
#> 245   A      BR2   25 0.59673942   Exponential -0.5162747507  0.74295485
#> 246   A      BR2   25 0.59673942 Monomolecular  0.9081723145  0.67823322
#> 247   A      BR2   25 0.59673942      Logistic  0.3918975638  0.69779258
#> 248   A      BR2   25 0.59673942      Gompertz  0.6611161926  0.68222018
#> 249   A      BR2   30 0.83481780   Exponential -0.1805417809  1.43273731
#> 250   A      BR2   30 0.83481780 Monomolecular  1.8007061749  0.75103383
#> 251   A      BR2   30 0.83481780      Logistic  1.6201643940  0.85195597
#> 252   A      BR2   30 0.83481780      Gompertz  1.7117930547  0.79007599
#> 253   A      BR2    0 0.01365471   Exponential -4.2936709604  0.02785747
#> 254   A      BR2    0 0.01365471 Monomolecular  0.0137487902 -0.16023391
#> 255   A      BR2    0 0.01365471      Logistic -4.2799221701  0.02344725
#> 256   A      BR2    0 0.01365471      Gompertz -1.4571420688  0.01349607
#> 257   A      BR2    5 0.07777727   Exponential -2.5539061009  0.05372122
#> 258   A      BR2    5 0.07777727 Monomolecular  0.0809685076  0.10227218
#> 259   A      BR2    5 0.07777727      Logistic -2.4729375933  0.05646253
#> 260   A      BR2    5 0.07777727      Gompertz -0.9376239914  0.07045064
#> 261   A      BR2   10 0.16085410   Exponential -1.8272575403  0.10359767
#> 262   A      BR2   10 0.16085410 Monomolecular  0.1753706893  0.30538555
#> 263   A      BR2   10 0.16085410      Logistic -1.6518868510  0.12978714
#> 264   A      BR2   10 0.16085410      Gompertz -0.6028162309  0.19502858
#> 265   A      BR2   15 0.53159190   Exponential -0.6318791910  0.19978098
#> 266   A      BR2   15 0.53159190 Monomolecular  0.7584153525  0.46254397
#> 267   A      BR2   15 0.53159190      Logistic  0.1265361615  0.27098609
#> 268   A      BR2   15 0.53159190      Gompertz  0.4590570565  0.36523971
#> 269   A      BR2   20 0.58468307   Exponential -0.5366853326  0.38526387
#> 270   A      BR2   20 0.58468307 Monomolecular  0.8787133728  0.58414486
#> 271   A      BR2   20 0.58468307      Logistic  0.3420280401  0.48090716
#> 272   A      BR2   20 0.58468307      Gompertz  0.6223433290  0.53761644
#> 273   A      BR2   25 0.61822535   Exponential -0.4809022366  0.74295485
#> 274   A      BR2   25 0.61822535 Monomolecular  0.9629247779  0.67823322
#> 275   A      BR2   25 0.61822535      Logistic  0.4820225413  0.69779258
#> 276   A      BR2   25 0.61822535      Gompertz  0.7320912799  0.68222018
#> 277   A      BR2   30 0.61471972   Exponential -0.4865888494  1.43273731
#> 278   A      BR2   30 0.61471972 Monomolecular  0.9537842184  0.75103383
#> 279   A      BR2   30 0.61471972      Logistic  0.4671953690  0.85195597
#> 280   A      BR2   30 0.61471972      Gompertz  0.7203357643  0.79007599
#> 281   A      BR2    0 0.01556841   Exponential -4.1625115434  0.02785747
#> 282   A      BR2    0 0.01556841 Monomolecular  0.0156908684 -0.16023391
#> 283   A      BR2    0 0.01556841      Logistic -4.1468206750  0.02344725
#> 284   A      BR2    0 0.01556841      Gompertz -1.4261186285  0.01349607
#> 285   A      BR2    5 0.05993785   Exponential -2.8144471154  0.05372122
#> 286   A      BR2    5 0.05993785 Monomolecular  0.0618092871  0.10227218
#> 287   A      BR2    5 0.05993785      Logistic -2.7526378283  0.05646253
#> 288   A      BR2    5 0.05993785      Gompertz -1.0347658357  0.07045064
#> 289   A      BR2   10 0.18254254   Exponential -1.7007720328  0.10359767
#> 290   A      BR2   10 0.18254254 Monomolecular  0.2015564153  0.30538555
#> 291   A      BR2   10 0.18254254      Logistic -1.4992156175  0.12978714
#> 292   A      BR2   10 0.18254254      Gompertz -0.5310822849  0.19502858
#> 293   A      BR2   15 0.46633403   Exponential -0.7628530984  0.19978098
#> 294   A      BR2   15 0.46633403 Monomolecular  0.6279851607  0.46254397
#> 295   A      BR2   15 0.46633403      Logistic -0.1348679377  0.27098609
#> 296   A      BR2   15 0.46633403      Gompertz  0.2706897979  0.36523971
#> 297   A      BR2   20 0.63832254   Exponential -0.4489115703  0.38526387
#> 298   A      BR2   20 0.63832254 Monomolecular  1.0170024668  0.58414486
#> 299   A      BR2   20 0.63832254      Logistic  0.5680908965  0.48090716
#> 300   A      BR2   20 0.63832254      Gompertz  0.8009293588  0.53761644
#> 301   A      BR2   25 0.74927365   Exponential -0.2886510037  0.74295485
#> 302   A      BR2   25 0.74927365 Monomolecular  1.3833931877  0.67823322
#> 303   A      BR2   25 0.74927365      Logistic  1.0947421840  0.69779258
#> 304   A      BR2   25 0.74927365      Gompertz  1.2425369203  0.68222018
#> 305   A      BR2   30 0.86996896   Exponential -0.1392977504  1.43273731
#> 306   A      BR2   30 0.86996896 Monomolecular  2.0399820590  0.75103383
#> 307   A      BR2   30 0.86996896      Logistic  1.9006843086  0.85195597
#> 308   A      BR2   30 0.86996896      Gompertz  1.9711415474  0.79007599
#> 309   A      BR2    0 0.01387157   Exponential -4.2779135191  0.02785747
#> 310   A      BR2    0 0.01387157 Monomolecular  0.0139686841 -0.16023391
#> 311   A      BR2    0 0.01387157      Logistic -4.2639448350  0.02344725
#> 312   A      BR2    0 0.01387157      Gompertz -1.4534653952  0.01349607
#> 313   A      BR2    5 0.06337225   Exponential -2.7587291905  0.05372122
#> 314   A      BR2    5 0.06337225 Monomolecular  0.0654693556  0.10227218
#> 315   A      BR2    5 0.06337225      Logistic -2.6932598349  0.05646253
#> 316   A      BR2    5 0.06337225      Gompertz -1.0147701355  0.07045064
#> 317   A      BR2   10 0.03456511   Exponential -3.3649105781  0.10359767
#> 318   A      BR2   10 0.03456511 Monomolecular  0.0351766127  0.30538555
#> 319   A      BR2   10 0.03456511      Logistic -3.3297339654  0.12978714
#> 320   A      BR2   10 0.03456511      Gompertz -1.2134013886  0.19502858
#> 321   A      BR2   15 0.42910466   Exponential -0.8460544234  0.19978098
#> 322   A      BR2   15 0.42910466 Monomolecular  0.5605493814  0.46254397
#> 323   A      BR2   15 0.42910466      Logistic -0.2855050420  0.27098609
#> 324   A      BR2   15 0.42910466      Gompertz  0.1671715912  0.36523971
#> 325   A      BR2   20 0.52922892   Exponential -0.6363342010  0.38526387
#> 326   A      BR2   20 0.52922892 Monomolecular  0.7533833314  0.58414486
#> 327   A      BR2   20 0.52922892      Logistic  0.1170491304  0.48090716
#> 328   A      BR2   20 0.52922892      Gompertz  0.4520313804  0.53761644
#> 329   A      BR2   25 0.66842178   Exponential -0.4028358910  0.74295485
#> 330   A      BR2   25 0.66842178 Monomolecular  1.1038915526  0.67823322
#> 331   A      BR2   25 0.66842178      Logistic  0.7010556617  0.69779258
#> 332   A      BR2   25 0.66842178      Gompertz  0.9092260184  0.68222018
#> 333   A      BR2   30 0.67531691   Exponential -0.3925732002  1.43273731
#> 334   A      BR2   30 0.67531691 Monomolecular  1.1249056839  0.75103383
#> 335   A      BR2   30 0.67531691      Logistic  0.7323324837  0.85195597
#> 336   A      BR2   30 0.67531691      Gompertz  0.9350322619  0.79007599
#> 337   B      BR2    0 0.01000000   Exponential -4.6051701860  0.02785747
#> 338   B      BR2    0 0.01000000 Monomolecular  0.0100503359 -0.16023391
#> 339   B      BR2    0 0.01000000      Logistic -4.5951198501  0.02344725
#> 340   B      BR2    0 0.01000000      Gompertz -1.5271796258  0.01349607
#> 341   B      BR2    5 0.06650718   Exponential -2.7104454417  0.05372122
#> 342   B      BR2    5 0.06650718 Monomolecular  0.0688220022  0.10227218
#> 343   B      BR2    5 0.06650718      Logistic -2.6416234395  0.05646253
#> 344   B      BR2    5 0.06650718      Gompertz -0.9971129910  0.07045064
#> 345   B      BR2   10 0.12108868   Exponential -2.1112320932  0.10359767
#> 346   B      BR2   10 0.12108868 Monomolecular  0.1290712760  0.30538555
#> 347   B      BR2   10 0.12108868      Logistic -1.9821608172  0.12978714
#> 348   B      BR2   10 0.12108868      Gompertz -0.7472717075  0.19502858
#> 349   B      BR2   15 0.54118615   Exponential -0.6139919688  0.19978098
#> 350   B      BR2   15 0.54118615 Monomolecular  0.7791107132  0.46254397
#> 351   B      BR2   15 0.54118615      Logistic  0.1651187445  0.27098609
#> 352   B      BR2   15 0.54118615      Gompertz  0.4877734311  0.36523971
#> 353   B      BR2   20 0.57718069   Exponential -0.5495999042  0.38526387
#> 354   B      BR2   20 0.57718069 Monomolecular  0.8608103584  0.58414486
#> 355   B      BR2   20 0.57718069      Logistic  0.3112104542  0.48090716
#> 356   B      BR2   20 0.57718069      Gompertz  0.5985647123  0.53761644
#> 357   B      BR2   25 0.59673942   Exponential -0.5162747507  0.74295485
#> 358   B      BR2   25 0.59673942 Monomolecular  0.9081723145  0.67823322
#> 359   B      BR2   25 0.59673942      Logistic  0.3918975638  0.69779258
#> 360   B      BR2   25 0.59673942      Gompertz  0.6611161926  0.68222018
#> 361   B      BR2   30 0.83481780   Exponential -0.1805417809  1.43273731
#> 362   B      BR2   30 0.83481780 Monomolecular  1.8007061749  0.75103383
#> 363   B      BR2   30 0.83481780      Logistic  1.6201643940  0.85195597
#> 364   B      BR2   30 0.83481780      Gompertz  1.7117930547  0.79007599
#> 365   B      BR2    0 0.01365471   Exponential -4.2936709604  0.02785747
#> 366   B      BR2    0 0.01365471 Monomolecular  0.0137487902 -0.16023391
#> 367   B      BR2    0 0.01365471      Logistic -4.2799221701  0.02344725
#> 368   B      BR2    0 0.01365471      Gompertz -1.4571420688  0.01349607
#> 369   B      BR2    5 0.07777727   Exponential -2.5539061009  0.05372122
#> 370   B      BR2    5 0.07777727 Monomolecular  0.0809685076  0.10227218
#> 371   B      BR2    5 0.07777727      Logistic -2.4729375933  0.05646253
#> 372   B      BR2    5 0.07777727      Gompertz -0.9376239914  0.07045064
#> 373   B      BR2   10 0.16085410   Exponential -1.8272575403  0.10359767
#> 374   B      BR2   10 0.16085410 Monomolecular  0.1753706893  0.30538555
#> 375   B      BR2   10 0.16085410      Logistic -1.6518868510  0.12978714
#> 376   B      BR2   10 0.16085410      Gompertz -0.6028162309  0.19502858
#> 377   B      BR2   15 0.53159190   Exponential -0.6318791910  0.19978098
#> 378   B      BR2   15 0.53159190 Monomolecular  0.7584153525  0.46254397
#> 379   B      BR2   15 0.53159190      Logistic  0.1265361615  0.27098609
#> 380   B      BR2   15 0.53159190      Gompertz  0.4590570565  0.36523971
#> 381   B      BR2   20 0.58468307   Exponential -0.5366853326  0.38526387
#> 382   B      BR2   20 0.58468307 Monomolecular  0.8787133728  0.58414486
#> 383   B      BR2   20 0.58468307      Logistic  0.3420280401  0.48090716
#> 384   B      BR2   20 0.58468307      Gompertz  0.6223433290  0.53761644
#> 385   B      BR2   25 0.61822535   Exponential -0.4809022366  0.74295485
#> 386   B      BR2   25 0.61822535 Monomolecular  0.9629247779  0.67823322
#> 387   B      BR2   25 0.61822535      Logistic  0.4820225413  0.69779258
#> 388   B      BR2   25 0.61822535      Gompertz  0.7320912799  0.68222018
#> 389   B      BR2   30 0.61471972   Exponential -0.4865888494  1.43273731
#> 390   B      BR2   30 0.61471972 Monomolecular  0.9537842184  0.75103383
#> 391   B      BR2   30 0.61471972      Logistic  0.4671953690  0.85195597
#> 392   B      BR2   30 0.61471972      Gompertz  0.7203357643  0.79007599
#> 393   B      BR2    0 0.01556841   Exponential -4.1625115434  0.02785747
#> 394   B      BR2    0 0.01556841 Monomolecular  0.0156908684 -0.16023391
#> 395   B      BR2    0 0.01556841      Logistic -4.1468206750  0.02344725
#> 396   B      BR2    0 0.01556841      Gompertz -1.4261186285  0.01349607
#> 397   B      BR2    5 0.05993785   Exponential -2.8144471154  0.05372122
#> 398   B      BR2    5 0.05993785 Monomolecular  0.0618092871  0.10227218
#> 399   B      BR2    5 0.05993785      Logistic -2.7526378283  0.05646253
#> 400   B      BR2    5 0.05993785      Gompertz -1.0347658357  0.07045064
#> 401   B      BR2   10 0.18254254   Exponential -1.7007720328  0.10359767
#> 402   B      BR2   10 0.18254254 Monomolecular  0.2015564153  0.30538555
#> 403   B      BR2   10 0.18254254      Logistic -1.4992156175  0.12978714
#> 404   B      BR2   10 0.18254254      Gompertz -0.5310822849  0.19502858
#> 405   B      BR2   15 0.46633403   Exponential -0.7628530984  0.19978098
#> 406   B      BR2   15 0.46633403 Monomolecular  0.6279851607  0.46254397
#> 407   B      BR2   15 0.46633403      Logistic -0.1348679377  0.27098609
#> 408   B      BR2   15 0.46633403      Gompertz  0.2706897979  0.36523971
#> 409   B      BR2   20 0.63832254   Exponential -0.4489115703  0.38526387
#> 410   B      BR2   20 0.63832254 Monomolecular  1.0170024668  0.58414486
#> 411   B      BR2   20 0.63832254      Logistic  0.5680908965  0.48090716
#> 412   B      BR2   20 0.63832254      Gompertz  0.8009293588  0.53761644
#> 413   B      BR2   25 0.74927365   Exponential -0.2886510037  0.74295485
#> 414   B      BR2   25 0.74927365 Monomolecular  1.3833931877  0.67823322
#> 415   B      BR2   25 0.74927365      Logistic  1.0947421840  0.69779258
#> 416   B      BR2   25 0.74927365      Gompertz  1.2425369203  0.68222018
#> 417   B      BR2   30 0.86996896   Exponential -0.1392977504  1.43273731
#> 418   B      BR2   30 0.86996896 Monomolecular  2.0399820590  0.75103383
#> 419   B      BR2   30 0.86996896      Logistic  1.9006843086  0.85195597
#> 420   B      BR2   30 0.86996896      Gompertz  1.9711415474  0.79007599
#> 421   B      BR2    0 0.01387157   Exponential -4.2779135191  0.02785747
#> 422   B      BR2    0 0.01387157 Monomolecular  0.0139686841 -0.16023391
#> 423   B      BR2    0 0.01387157      Logistic -4.2639448350  0.02344725
#> 424   B      BR2    0 0.01387157      Gompertz -1.4534653952  0.01349607
#> 425   B      BR2    5 0.06337225   Exponential -2.7587291905  0.05372122
#> 426   B      BR2    5 0.06337225 Monomolecular  0.0654693556  0.10227218
#> 427   B      BR2    5 0.06337225      Logistic -2.6932598349  0.05646253
#> 428   B      BR2    5 0.06337225      Gompertz -1.0147701355  0.07045064
#> 429   B      BR2   10 0.03456511   Exponential -3.3649105781  0.10359767
#> 430   B      BR2   10 0.03456511 Monomolecular  0.0351766127  0.30538555
#> 431   B      BR2   10 0.03456511      Logistic -3.3297339654  0.12978714
#> 432   B      BR2   10 0.03456511      Gompertz -1.2134013886  0.19502858
#> 433   B      BR2   15 0.42910466   Exponential -0.8460544234  0.19978098
#> 434   B      BR2   15 0.42910466 Monomolecular  0.5605493814  0.46254397
#> 435   B      BR2   15 0.42910466      Logistic -0.2855050420  0.27098609
#> 436   B      BR2   15 0.42910466      Gompertz  0.1671715912  0.36523971
#> 437   B      BR2   20 0.52922892   Exponential -0.6363342010  0.38526387
#> 438   B      BR2   20 0.52922892 Monomolecular  0.7533833314  0.58414486
#> 439   B      BR2   20 0.52922892      Logistic  0.1170491304  0.48090716
#> 440   B      BR2   20 0.52922892      Gompertz  0.4520313804  0.53761644
#> 441   B      BR2   25 0.66842178   Exponential -0.4028358910  0.74295485
#> 442   B      BR2   25 0.66842178 Monomolecular  1.1038915526  0.67823322
#> 443   B      BR2   25 0.66842178      Logistic  0.7010556617  0.69779258
#> 444   B      BR2   25 0.66842178      Gompertz  0.9092260184  0.68222018
#> 445   B      BR2   30 0.67531691   Exponential -0.3925732002  1.43273731
#> 446   B      BR2   30 0.67531691 Monomolecular  1.1249056839  0.75103383
#> 447   B      BR2   30 0.67531691      Logistic  0.7323324837  0.85195597
#> 448   B      BR2   30 0.67531691      Gompertz  0.9350322619  0.79007599
#>          residual
#> 1   -7.707858e-02
#> 2    6.576388e-01
#> 3   -4.019756e-02
#> 4   -8.387862e-03
#> 5    2.289218e-01
#> 6   -1.696796e-01
#> 7    1.294531e-01
#> 8   -2.747836e-02
#> 9    4.684024e-01
#> 10  -1.493228e-01
#> 11   5.044830e-02
#> 12  -8.938833e-02
#> 13   5.421564e-01
#> 14   2.178728e-02
#> 15   5.852758e-02
#> 16   3.258275e-02
#> 17   2.222407e-01
#> 18  -2.260997e-04
#> 19   2.427956e-03
#> 20   7.374869e-04
#> 21  -3.276382e-01
#> 22  -1.044159e-03
#> 23  -1.669705e-03
#> 24  -1.269081e-03
#> 25  -1.282370e+00
#> 26   2.659545e-04
#> 27  -1.251399e-04
#> 28   9.215872e-05
#> 29  -7.342387e-02
#> 30   6.612935e-01
#> 31  -3.654285e-02
#> 32  -4.733155e-03
#> 33   2.739797e-01
#> 34  -1.246217e-01
#> 35   1.745110e-01
#> 36   1.757957e-02
#> 37   5.115924e-01
#> 38  -1.061328e-01
#> 39   9.363822e-02
#> 40  -4.619841e-02
#> 41   5.401780e-01
#> 42   1.980887e-02
#> 43   5.654918e-02
#> 44   3.060435e-02
#> 45   2.225793e-01
#> 46   1.124494e-04
#> 47   2.766506e-03
#> 48   1.076036e-03
#> 49  -3.273854e-01
#> 50  -7.913796e-04
#> 51  -1.416926e-03
#> 52  -1.016301e-03
#> 53  -1.283138e+00
#> 54  -5.017378e-04
#> 55  -8.928321e-04
#> 56  -6.755335e-04
#> 57  -7.151017e-02
#> 58   6.632072e-01
#> 59  -3.462915e-02
#> 60  -2.819454e-03
#> 61   2.026576e-01
#> 62  -1.959438e-01
#> 63   1.031888e-01
#> 64  -5.374259e-02
#> 65   5.351486e-01
#> 66  -8.257665e-02
#> 67   1.171944e-01
#> 68  -2.264221e-02
#> 69   5.267213e-01
#> 70   6.352217e-03
#> 71   4.309252e-02
#> 72   1.714769e-02
#> 73   2.249998e-01
#> 74   2.532959e-03
#> 75   5.187016e-03
#> 76   3.496546e-03
#> 77  -3.258437e-01
#> 78   7.503866e-04
#> 79   1.248402e-04
#> 80   5.254647e-04
#> 81  -1.282247e+00
#> 82   3.885601e-04
#> 83  -2.534217e-06
#> 84   2.147644e-04
#> 85  -7.320701e-02
#> 86   6.615103e-01
#> 87  -3.632599e-02
#> 88  -4.516287e-03
#> 89   2.163883e-01
#> 90  -1.822131e-01
#> 91   1.169196e-01
#> 92  -4.001182e-02
#> 93   3.744277e-01
#> 94  -2.432976e-01
#> 95  -4.352649e-02
#> 96  -1.833631e-01
#> 97   5.190444e-01
#> 98  -1.324754e-03
#> 99   3.541555e-02
#> 100  9.470721e-03
#> 101  2.200769e-01
#> 102 -2.389949e-03
#> 103  2.641069e-04
#> 104 -1.426363e-03
#> 105 -3.267949e-01
#> 106 -2.008251e-04
#> 107 -8.263715e-04
#> 108 -4.257470e-04
#> 109 -1.282926e+00
#> 110 -2.903775e-04
#> 111 -6.814718e-04
#> 112 -4.641732e-04
#> 113 -4.971695e-02
#> 114  4.618776e-01
#> 115 -2.950593e-02
#> 116 -6.237192e-03
#> 117  8.828658e-02
#> 118 -1.308560e-01
#> 119  5.762601e-02
#> 120 -1.890881e-02
#> 121  2.324403e-01
#> 122 -2.570171e-01
#> 123  4.082265e-02
#> 124 -1.341048e-01
#> 125  5.593937e-01
#> 126  6.876515e-02
#> 127  2.071864e-01
#> 128  1.153817e-01
#> 129  2.623893e-01
#> 130 -2.162539e-03
#> 131  2.175014e-02
#> 132  6.780262e-03
#> 133 -2.665753e-01
#> 134 -1.195520e-02
#> 135 -1.819592e-02
#> 136 -1.400103e-02
#> 137 -1.247660e+00
#> 138  5.608034e-03
#> 139 -2.303675e-03
#> 140  2.142683e-03
#> 141 -4.606224e-02
#> 142  4.655323e-01
#> 143 -2.585123e-02
#> 144 -2.582485e-03
#> 145  1.176960e-01
#> 146 -1.014466e-01
#> 147  8.703543e-02
#> 148  1.050060e-02
#> 149  2.983726e-01
#> 150 -1.910848e-01
#> 151  1.067549e-01
#> 152 -6.817255e-02
#> 153  5.525924e-01
#> 154  6.196379e-02
#> 155  2.003851e-01
#> 156  1.085804e-01
#> 157  2.646321e-01
#> 158  8.028156e-05
#> 159  2.399296e-02
#> 160  9.023083e-03
#> 161 -2.636246e-01
#> 162 -9.004555e-03
#> 163 -1.524528e-02
#> 164 -1.105039e-02
#> 165 -1.262823e+00
#> 166 -9.554754e-03
#> 167 -1.746646e-02
#> 168 -1.302011e-02
#> 169 -4.414854e-02
#> 170  4.674460e-01
#> 171 -2.393753e-02
#> 172 -6.687839e-04
#> 173  7.114386e-02
#> 174 -1.479988e-01
#> 175  4.048329e-02
#> 176 -3.605154e-02
#> 177  3.343327e-01
#> 178 -1.551247e-01
#> 179  1.427150e-01
#> 180 -3.221247e-02
#> 181  5.063311e-01
#> 182  1.570254e-02
#> 183  1.541238e-01
#> 184  6.231912e-02
#> 185  2.806675e-01
#> 186  1.611568e-02
#> 187  4.002836e-02
#> 188  2.505848e-02
#> 189 -2.456279e-01
#> 190  8.992189e-03
#> 191  2.751469e-03
#> 192  6.946357e-03
#> 193 -1.245239e+00
#> 194  8.029634e-03
#> 195  1.179251e-04
#> 196  4.564284e-03
#> 197 -4.584538e-02
#> 198  4.657491e-01
#> 199 -2.563436e-02
#> 200 -2.365617e-03
#> 201  8.010597e-02
#> 202 -1.390367e-01
#> 203  4.944539e-02
#> 204 -2.708943e-02
#> 205  8.898163e-02
#> 206 -4.004758e-01
#> 207 -1.026360e-01
#> 208 -2.775635e-01
#> 209  4.799392e-01
#> 210 -1.068934e-02
#> 211  1.277319e-01
#> 212  3.592724e-02
#> 213  2.480542e-01
#> 214 -1.649762e-02
#> 215  7.415059e-03
#> 216 -7.554816e-03
#> 217 -2.567312e-01
#> 218 -2.111125e-03
#> 219 -8.351845e-03
#> 220 -4.156957e-03
#> 221 -1.258648e+00
#> 222 -5.380150e-03
#> 223 -1.329186e-02
#> 224 -8.845501e-03
#> 225 -1.785747e-02
#> 226  1.702339e-01
#> 227 -1.344725e-02
#> 228 -3.496069e-03
#> 229  1.278596e-02
#> 230 -3.576500e-02
#> 231  1.004465e-02
#> 232 -3.943465e-03
#> 233  1.749101e-02
#> 234 -1.842969e-01
#> 235 -8.698461e-03
#> 236 -7.393990e-02
#> 237  3.414052e-01
#> 238  7.864219e-02
#> 239  2.702001e-01
#> 240  1.759464e-01
#> 241  1.919168e-01
#> 242 -6.964169e-03
#> 243  9.627353e-02
#> 244  3.956425e-02
#> 245 -1.462154e-01
#> 246 -8.149381e-02
#> 247 -1.010532e-01
#> 248 -8.548076e-02
#> 249 -5.979195e-01
#> 250  8.378397e-02
#> 251 -1.713817e-02
#> 252  4.474181e-02
#> 253 -1.420276e-02
#> 254  1.738886e-01
#> 255 -9.792539e-03
#> 256  1.586386e-04
#> 257  2.405605e-02
#> 258 -2.449491e-02
#> 259  2.131474e-02
#> 260  7.326626e-03
#> 261  5.725643e-02
#> 262 -1.445314e-01
#> 263  3.106696e-02
#> 264 -3.417448e-02
#> 265  3.318109e-01
#> 266  6.904793e-02
#> 267  2.606058e-01
#> 268  1.663522e-01
#> 269  1.994192e-01
#> 270  5.382126e-04
#> 271  1.037759e-01
#> 272  4.706664e-02
#> 273 -1.247295e-01
#> 274 -6.000787e-02
#> 275 -7.956723e-02
#> 276 -6.399482e-02
#> 277 -8.180176e-01
#> 278 -1.363141e-01
#> 279 -2.372362e-01
#> 280 -1.753563e-01
#> 281 -1.228906e-02
#> 282  1.758023e-01
#> 283 -7.878839e-03
#> 284  2.072339e-03
#> 285  6.216631e-03
#> 286 -4.233433e-02
#> 287  3.475321e-03
#> 288 -1.051279e-02
#> 289  7.894487e-02
#> 290 -1.228430e-01
#> 291  5.275540e-02
#> 292 -1.248604e-02
#> 293  2.665531e-01
#> 294  3.790065e-03
#> 295  1.953479e-01
#> 296  1.010943e-01
#> 297  2.530587e-01
#> 298  5.417768e-02
#> 299  1.574154e-01
#> 300  1.007061e-01
#> 301  6.318799e-03
#> 302  7.104043e-02
#> 303  5.148107e-02
#> 304  6.705347e-02
#> 305 -5.627684e-01
#> 306  1.189351e-01
#> 307  1.801299e-02
#> 308  7.989297e-02
#> 309 -1.398590e-02
#> 310  1.741055e-01
#> 311 -9.575672e-03
#> 312  3.755060e-04
#> 313  9.651034e-03
#> 314 -3.889993e-02
#> 315  6.909724e-03
#> 316 -7.078388e-03
#> 317 -6.903256e-02
#> 318 -2.708204e-01
#> 319 -9.522204e-02
#> 320 -1.604635e-01
#> 321  2.293237e-01
#> 322 -3.343930e-02
#> 323  1.581186e-01
#> 324  6.386495e-02
#> 325  1.439651e-01
#> 326 -5.491594e-02
#> 327  4.832176e-02
#> 328 -8.387518e-03
#> 329 -7.453307e-02
#> 330 -9.811438e-03
#> 331 -2.937080e-02
#> 332 -1.379839e-02
#> 333 -7.574204e-01
#> 334 -7.571692e-02
#> 335 -1.766391e-01
#> 336 -1.147591e-01
#> 337 -1.785747e-02
#> 338  1.702339e-01
#> 339 -1.344725e-02
#> 340 -3.496069e-03
#> 341  1.278596e-02
#> 342 -3.576500e-02
#> 343  1.004465e-02
#> 344 -3.943465e-03
#> 345  1.749101e-02
#> 346 -1.842969e-01
#> 347 -8.698461e-03
#> 348 -7.393990e-02
#> 349  3.414052e-01
#> 350  7.864219e-02
#> 351  2.702001e-01
#> 352  1.759464e-01
#> 353  1.919168e-01
#> 354 -6.964169e-03
#> 355  9.627353e-02
#> 356  3.956425e-02
#> 357 -1.462154e-01
#> 358 -8.149381e-02
#> 359 -1.010532e-01
#> 360 -8.548076e-02
#> 361 -5.979195e-01
#> 362  8.378397e-02
#> 363 -1.713817e-02
#> 364  4.474181e-02
#> 365 -1.420276e-02
#> 366  1.738886e-01
#> 367 -9.792539e-03
#> 368  1.586386e-04
#> 369  2.405605e-02
#> 370 -2.449491e-02
#> 371  2.131474e-02
#> 372  7.326626e-03
#> 373  5.725643e-02
#> 374 -1.445314e-01
#> 375  3.106696e-02
#> 376 -3.417448e-02
#> 377  3.318109e-01
#> 378  6.904793e-02
#> 379  2.606058e-01
#> 380  1.663522e-01
#> 381  1.994192e-01
#> 382  5.382126e-04
#> 383  1.037759e-01
#> 384  4.706664e-02
#> 385 -1.247295e-01
#> 386 -6.000787e-02
#> 387 -7.956723e-02
#> 388 -6.399482e-02
#> 389 -8.180176e-01
#> 390 -1.363141e-01
#> 391 -2.372362e-01
#> 392 -1.753563e-01
#> 393 -1.228906e-02
#> 394  1.758023e-01
#> 395 -7.878839e-03
#> 396  2.072339e-03
#> 397  6.216631e-03
#> 398 -4.233433e-02
#> 399  3.475321e-03
#> 400 -1.051279e-02
#> 401  7.894487e-02
#> 402 -1.228430e-01
#> 403  5.275540e-02
#> 404 -1.248604e-02
#> 405  2.665531e-01
#> 406  3.790065e-03
#> 407  1.953479e-01
#> 408  1.010943e-01
#> 409  2.530587e-01
#> 410  5.417768e-02
#> 411  1.574154e-01
#> 412  1.007061e-01
#> 413  6.318799e-03
#> 414  7.104043e-02
#> 415  5.148107e-02
#> 416  6.705347e-02
#> 417 -5.627684e-01
#> 418  1.189351e-01
#> 419  1.801299e-02
#> 420  7.989297e-02
#> 421 -1.398590e-02
#> 422  1.741055e-01
#> 423 -9.575672e-03
#> 424  3.755060e-04
#> 425  9.651034e-03
#> 426 -3.889993e-02
#> 427  6.909724e-03
#> 428 -7.078388e-03
#> 429 -6.903256e-02
#> 430 -2.708204e-01
#> 431 -9.522204e-02
#> 432 -1.604635e-01
#> 433  2.293237e-01
#> 434 -3.343930e-02
#> 435  1.581186e-01
#> 436  6.386495e-02
#> 437  1.439651e-01
#> 438 -5.491594e-02
#> 439  4.832176e-02
#> 440 -8.387518e-03
#> 441 -7.453307e-02
#> 442 -9.811438e-03
#> 443 -2.937080e-02
#> 444 -1.379839e-02
#> 445 -7.574204e-01
#> 446 -7.571692e-02
#> 447 -1.766391e-01
#> 448 -1.147591e-01
#>